--- src/main/java/com/yzgeneration/evc/domain/image/controller/ImageController.java
@@ -1,22 +1,0 @@
-package com.yzgeneration.evc.domain.image.controller;
-
-import com.yzgeneration.evc.domain.image.dto.ImageResponse;
-import com.yzgeneration.evc.domain.image.service.ImageService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/images")
-@RequiredArgsConstructor
-public class ImageController {
-    private final ImageService imageService;
-
-    @PostMapping(consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    public List<ImageResponse> createPresignedURL(@RequestParam String prefix, @RequestPart List<MultipartFile> imageFiles) {
-        return imageService.generatePresignedURL(prefix, imageFiles);
-    }
-}
